requirements.txt
flask
mysql-connector-python
python-dotenv
werkzeug
openai
PyPDF2
python-docx
openpyxl
xlrd
sumy
transformers 
torch